Hiresmart/Hiresmart/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.4.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
flasgger==0.9.7.1
Werkzeug==2.3.7
Flask-Mail==0.9.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
